Add documentation, move some functions (which are not changed)
