+++ src/common/util/util.ts
@@ -1,5 +1,3 @@
-/** @module common/util/util */
-
 import { timeout } from './timeout.js';
 
 /**
