+++ src/webgpu/api/validation/image_copy/image_copy.ts
@@ -1,5 +1,5 @@
 import { kSizedTextureFormatInfo, SizedTextureFormat } from '../../../capability_info.js';
-import { ImageCopyType } from '../../../util/texture/image_copy.js';
+import { ImageCopyType } from '../../../util/texture/layout.js';
 import { ValidationTest } from '../validation_test.js';
 
 export class ImageCopyTest extends ValidationTest {
