+++ src/webgpu/api/validation/validation_test.ts
@@ -24,6 +24,10 @@ export interface CommandBufferMaker<T extends EncoderType> {
  * Base fixture for WebGPU validation tests.
  */
 export class ValidationTest extends GPUTest {
+  /**
+   * Create a GPUTexture in the specified state.
+   * A `descriptor` may optionally be passed, which is used when `state` is not `'invalid'`.
+   */
   createTextureWithState(
     state: 'valid' | 'invalid' | 'destroyed',
     descriptor?: Readonly<GPUTextureDescriptor>
@@ -52,6 +56,10 @@ export class ValidationTest extends GPUTest {
     }
   }
 
+  /**
+   * Create a GPUTexture in the specified state. A `descriptor` may optionally be passed;
+   * if `state` is `'invalid'`, it will be modified to add an invalid combination of usages.
+   */
   createBufferWithState(
     state: 'valid' | 'invalid' | 'destroyed',
     descriptor?: Readonly<GPUBufferDescriptor>
@@ -84,6 +92,10 @@ export class ValidationTest extends GPUTest {
     }
   }
 
+  /**
+   * Create a GPUQuerySet in the specified state.
+   * A `descriptor` may optionally be passed, which is used when `state` is not `'invalid'`.
+   */
   createQuerySetWithState(
     state: 'valid' | 'invalid' | 'destroyed',
     descriptor?: Readonly<GPUQuerySetDescriptor>
@@ -114,18 +126,22 @@ export class ValidationTest extends GPUTest {
     }
   }
 
+  /** Create an arbitrarily-sized GPUBuffer with the STORAGE usage. */
   getStorageBuffer(): GPUBuffer {
     return this.device.createBuffer({ size: 1024, usage: GPUBufferUsage.STORAGE });
   }
 
+  /** Create an arbitrarily-sized GPUBuffer with the UNIFORM usage. */
   getUniformBuffer(): GPUBuffer {
     return this.device.createBuffer({ size: 1024, usage: GPUBufferUsage.UNIFORM });
   }
 
+  /** Return an invalid GPUBuffer. */
   getErrorBuffer(): GPUBuffer {
     return this.createBufferWithState('invalid');
   }
 
+  /** Return an invalid GPUSampler. */
   getErrorSampler(): GPUSampler {
     this.device.pushErrorScope('validation');
     const sampler = this.device.createSampler({ lodMinClamp: -1 });
@@ -133,6 +149,9 @@ export class ValidationTest extends GPUTest {
     return sampler;
   }
 
+  /**
+   * Return an arbitrarily-configured GPUTexture with the `SAMPLED` usage and specified sampleCount.
+   */
   getSampledTexture(sampleCount: number = 1): GPUTexture {
     return this.device.createTexture({
       size: { width: 16, height: 16, depthOrArrayLayers: 1 },
@@ -142,6 +161,7 @@ export class ValidationTest extends GPUTest {
     });
   }
 
+  /** Return an arbitrarily-configured GPUTexture with the `STORAGE` usage. */
   getStorageTexture(): GPUTexture {
     return this.device.createTexture({
       size: { width: 16, height: 16, depthOrArrayLayers: 1 },
@@ -150,6 +170,7 @@ export class ValidationTest extends GPUTest {
     });
   }
 
+  /** Return an arbitrarily-configured GPUTexture with the `RENDER_ATTACHMENT` usage. */
   getRenderTexture(): GPUTexture {
     return this.device.createTexture({
       size: { width: 16, height: 16, depthOrArrayLayers: 1 },
@@ -158,6 +179,7 @@ export class ValidationTest extends GPUTest {
     });
   }
 
+  /** Return an invalid GPUTexture. */
   getErrorTexture(): GPUTexture {
     this.device.pushErrorScope('validation');
     const texture = this.device.createTexture({
@@ -169,6 +191,7 @@ export class ValidationTest extends GPUTest {
     return texture;
   }
 
+  /** Return an invalid GPUTextureView (created from an invalid GPUTexture). */
   getErrorTextureView(): GPUTextureView {
     this.device.pushErrorScope('validation');
     const view = this.getErrorTexture().createView();
@@ -176,6 +199,10 @@ export class ValidationTest extends GPUTest {
     return view;
   }
 
+  /**
+   * Return an arbitrary object of the specified {@link BindableResource} type
+   * (e.g. `'errorBuf'`, `'nonFiltSamp'`, `sampledTexMS`, etc.)
+   */
   getBindingResource(bindingType: BindableResource): GPUBindingResource {
     switch (bindingType) {
       case 'errorBuf':
@@ -203,6 +230,7 @@ export class ValidationTest extends GPUTest {
     }
   }
 
+  /** Return a GPURenderPipeline with default options and no-op vertex and fragment shaders. */
   createNoOpRenderPipeline(): GPURenderPipeline {
     return this.device.createRenderPipeline({
       vertex: {
@@ -222,6 +250,7 @@ export class ValidationTest extends GPUTest {
     });
   }
 
+  /** Return a GPUComputePipeline with a no-op shader. */
   createNoOpComputePipeline(): GPUComputePipeline {
     return this.device.createComputePipeline({
       compute: {
@@ -233,6 +262,7 @@ export class ValidationTest extends GPUTest {
     });
   }
 
+  /** Return an invalid GPUComputePipeline. */
   createErrorComputePipeline(): GPUComputePipeline {
     this.device.pushErrorScope('validation');
     const pipeline = this.device.createComputePipeline({
@@ -247,6 +277,33 @@ export class ValidationTest extends GPUTest {
     return pipeline;
   }
 
+  /**
+   * Returns a GPUCommandEncoder, GPUComputePassEncoder, GPURenderPassEncoder, or
+   * GPURenderBundleEncoder, and a `finish` method returning a GPUCommandBuffer.
+   * Allows testing methods which have the same signature across multiple encoder interfaces.
+   *
+   * TODO(https://github.com/gpuweb/cts/pull/489#issuecomment-812283347):
+   * Make this have stricter validation to ensure errors are generated in the right API call.
+   *
+   * @example
+   * ```
+   * g.test('popDebugGroup')
+   *   .params(u => u.combine('encoderType', kEncoderTypes))
+   *   .fn(t => {
+   *     const { encoder, finish } = t.createEncoder(t.params.encoderType);
+   *     encoder.popDebugGroup();
+   *   });
+   *
+   * g.test('writeTimestamp')
+   *   .params(u => u.combine('encoderType', ['non-pass', 'compute pass', 'render pass'] as const)
+   *   .fn(t => {
+   *     const { encoder, finish } = t.createEncoder(t.params.encoderType);
+   *     // Encoder type is inferred, so `writeTimestamp` can be used even though it doesn't exist
+   *     // on GPURenderBundleEncoder.
+   *     encoder.writeTimestamp(args);
+   *   });
+   * ```
+   */
   createEncoder<T extends EncoderType>(encoderType: T): CommandBufferMaker<T> {
     const colorFormat = 'rgba8unorm';
     switch (encoderType) {
@@ -320,21 +377,22 @@ export class ValidationTest extends GPUTest {
    *
    * Tests should always do just one WebGPU call in the callback, to make sure that's what's tested.
    */
-  // Note: A return value is not allowed for the callback function. This is to avoid confusion
-  // about what the actual behavior would be. We could either:
-  //   - Make expectValidationError async, and have it await on fn(). This causes an async split
-  //     between pushErrorScope and popErrorScope, so if the caller doesn't `await` on
-  //     expectValidationError (either accidentally or because it doesn't care to do so), then
-  //     other test code will be (nondeterministically) caught by the error scope.
-  //   - Make expectValidationError NOT await fn(), but just execute its first block (until the
-  //     first await) and return the return value (a Promise). This would be confusing because it
-  //     would look like the error scope includes the whole async function, but doesn't.
   expectValidationError(fn: () => void, shouldError: boolean = true): void {
     // If no error is expected, we let the scope surrounding the test catch it.
     if (shouldError) {
       this.device.pushErrorScope('validation');
     }
 
+    // Note: A return value is not allowed for the callback function. This is to avoid confusion
+    // about what the actual behavior would be; either of the following could be reasonable:
+    //   - Make expectValidationError async, and have it await on fn(). This causes an async split
+    //     between pushErrorScope and popErrorScope, so if the caller doesn't `await` on
+    //     expectValidationError (either accidentally or because it doesn't care to do so), then
+    //     other test code will be (nondeterministically) caught by the error scope.
+    //   - Make expectValidationError NOT await fn(), but just execute its first block (until the
+    //     first await) and return the return value (a Promise). This would be confusing because it
+    //     would look like the error scope includes the whole async function, but doesn't.
+    // If we do decide we need to return a value, we should use the latter semantic.
     const returnValue = fn() as unknown;
     assert(
       returnValue === undefined,
