+++ src/webgpu/constants.ts
@@ -1,5 +1,3 @@
-/** @module webgpu/constants */
-
 // Note: Types ensure every field is specified.
 
 /* eslint-disable-next-line @typescript-eslint/no-unused-vars */
