+++ src/webgpu/util/conversion.ts
@@ -2,6 +2,10 @@ import { assert } from '../../common/util/util.js';
 
 import { clamp } from './math.js';
 
+/**
+ * Encodes a JS `number` into a "normalized" (unorm/snorm) integer representation with `bits` bits.
+ * Input must be between -1 and 1 if signed, or 0 and 1 if unsigned.
+ */
 export function floatAsNormalizedInteger(float: number, bits: number, signed: boolean): number {
   if (signed) {
     assert(float >= -1 && float <= 1);
@@ -14,6 +18,10 @@ export function floatAsNormalizedInteger(float: number, bits: number, signed: bo
   }
 }
 
+/**
+ * Decodes a JS `number` from a "normalized" (unorm/snorm) integer representation with `bits` bits.
+ * Input must be an integer in the range of the specified unorm/snorm type.
+ */
 export function normalizedIntegerAsFloat(integer: number, bits: number, signed: boolean): number {
   assert(Number.isInteger(integer));
   if (signed) {
@@ -30,7 +38,13 @@ export function normalizedIntegerAsFloat(integer: number, bits: number, signed: 
   }
 }
 
-// Does not handle clamping, underflow, overflow, denormalized numbers
+/**
+ * Encodes a JS `number` into an IEEE754 floating point number with the specified number of
+ * sign, exponent, mantissa bits, and exponent bias.
+ * Returns the result as an integer-valued JS `number`.
+ *
+ * Does not handle clamping, underflow, overflow, or denormalized numbers.
+ */
 export function float32ToFloatBits(
   n: number,
   signBits: 0 | 1,
@@ -72,11 +86,15 @@ export function float32ToFloatBits(
   return (sign << (exponentBits + mantissaBits)) | (newBiasedExp << mantissaBits) | newMantissa;
 }
 
-// Three partial-precision floating-point numbers encoded into a single 32-bit value all
-// sharing the same 5-bit exponent.
-// There is no sign bit, and there is a shared 5-bit biased (15) exponent and a 9-bit
-// mantissa for each channel. The mantissa does NOT have an implicit leading "1.",
-// and instead has an implicit leading "0.".
+/**
+ * Encodes three JS `number` values into RGB9E5, returned as an integer-valued JS `number`.
+ *
+ * RGB9E5 represents three partial-precision floating-point numbers encoded into a single 32-bit
+ * value all sharing the same 5-bit exponent.
+ * There is no sign bit, and there is a shared 5-bit biased (15) exponent and a 9-bit
+ * mantissa for each channel. The mantissa does NOT have an implicit leading "1.",
+ * and instead has an implicit leading "0.".
+ */
 export function packRGB9E5UFloat(r: number, g: number, b: number): number {
   for (const v of [r, g, b]) {
     assert(v >= 0 && v < Math.pow(2, 16));
@@ -119,6 +137,12 @@ export function packRGB9E5UFloat(r: number, g: number, b: number): number {
   return rMantissa | (gMantissa << 9) | (bMantissa << 18) | (biasedExp << 27);
 }
 
+/**
+ * Asserts that a number is within the representable (inclusive) of the integer type with the
+ * specified number of bits and signedness.
+ *
+ * TODO: Assert isInteger? Then this function "asserts that a number is representable" by the type
+ */
 export function assertInIntegerRange(n: number, bits: number, signed: boolean): void {
   if (signed) {
     const min = -Math.pow(2, bits - 1);
@@ -130,12 +154,18 @@ export function assertInIntegerRange(n: number, bits: number, signed: boolean): 
   }
 }
 
+/**
+ * Converts a linear value into a "gamma"-encoded value using the sRGB-clamped transfer function.
+ */
 export function gammaCompress(n: number): number {
   n = n <= 0.0031308 ? (323 * n) / 25 : (211 * Math.pow(n, 5 / 12) - 11) / 200;
-  return clamp(n, 0, 1);
+  return clamp(n, { min: 0, max: 1 });
 }
 
+/**
+ * Converts a "gamma"-encoded value into a linear value using the sRGB-clamped transfer function.
+ */
 export function gammaDecompress(n: number): number {
   n = n <= 0.04045 ? (n * 25) / 323 : Math.pow((200 * n + 11) / 211, 12 / 5);
-  return clamp(n, 0, 1);
+  return clamp(n, { min: 0, max: 1 });
 }
