+++ src/webgpu/util/math.ts
@@ -1,16 +1,23 @@
 import { assert } from '../../common/util/util.js';
 
+/**
+ * A multiple of 8 guaranteed to be way too large to allocate (just under 8 pebibytes).
+ * This is a "safe" integer (ULP <= 1.0) very close to MAX_SAFE_INTEGER.
+ *
+ * Note: allocations of this size are likely to exceed limitations other than just the system's
+ * physical memory, so test cases are also needed to try to trigger "true" OOM.
+ */
+export const kMaxSafeMultipleOf8 = Number.MAX_SAFE_INTEGER - 7;
+
+/** Round `n` up to the next multiple of `alignment`. */
 export function align(n: number, alignment: number): number {
+  assert(Number.isInteger(n) && n >= 0, 'n must be a non-negative integer');
+  assert(Number.isInteger(alignment) && alignment > 0, 'alignment must be a positive integer');
   return Math.ceil(n / alignment) * alignment;
 }
 
-export function isAligned(n: number, alignment: number): boolean {
-  return n === align(n, alignment);
-}
-
-export const kMaxSafeMultipleOf8 = Number.MAX_SAFE_INTEGER - 7;
-
-export function clamp(n: number, min: number, max: number): number {
+/** Clamp a number to the provided range. */
+export function clamp(n: number, { min, max }: { min: number; max: number }): number {
   assert(max >= min);
   return Math.min(Math.max(n, min), max);
 }
