+++ src/webgpu/util/navigator_gpu.ts
@@ -4,6 +4,10 @@ import { assert } from '../../common/util/util.js';
 
 let impl: GPU | undefined = undefined;
 
+/**
+ * Finds and returns the `navigator.gpu` object (or equivalent, for non-browser implementations).
+ * Throws an exception if not found.
+ */
 export function getGPU(): GPU {
   if (impl) {
     return impl;
