+++ src/webgpu/util/texture/layout.ts
@@ -1,12 +1,18 @@
-import { assert, unreachable } from '../../../common/util/util.js';
+import { assert } from '../../../common/util/util.js';
 import { kSizedTextureFormatInfo, SizedTextureFormat } from '../../capability_info.js';
-import { align, isAligned } from '../math.js';
+import { align } from '../math.js';
+import { standardizeExtent3D } from '../unions.js';
 
-import { bytesInACompleteRow } from './image_copy.js';
+import { getMipSizePassthroughLayers } from './base.js';
 
+/** The minimum `bytesPerRow` alignment, per spec. */
 export const kBytesPerRowAlignment = 256;
+/** The minimum buffer copy alignment, per spec. */
 export const kBufferCopyAlignment = 4;
 
+/**
+ * Overridable layout options for {@link getTextureCopyLayout}.
+ */
 export interface LayoutOptions {
   mipLevel: number;
   bytesPerRow?: number;
@@ -15,34 +21,29 @@ export interface LayoutOptions {
 
 const kDefaultLayoutOptions = { mipLevel: 0, bytesPerRow: undefined, rowsPerImage: undefined };
 
-export function getMipSizePassthroughLayers(
-  dimension: GPUTextureDimension,
-  size: readonly [number, number, number],
-  mipLevel: number
-): [number, number, number] {
-  const shiftMinOne = (n: number) => Math.max(1, n >> mipLevel);
-  switch (dimension) {
-    case '1d':
-      assert(size[2] === 1);
-      return [shiftMinOne(size[0]), size[1], size[2]];
-    case '2d':
-      return [shiftMinOne(size[0]), shiftMinOne(size[1]), size[2]];
-    case '3d':
-      return [shiftMinOne(size[0]), shiftMinOne(size[1]), shiftMinOne(size[2])];
-    default:
-      unreachable();
-  }
-}
-
+/** The info returned by {@link getTextureCopyLayout}. */
 export interface TextureCopyLayout {
   bytesPerBlock: number;
   byteLength: number;
+  /** Number of bytes in each row, not accounting for {@link kBytesPerRowAlignment}. */
   minBytesPerRow: number;
+  /**
+   * Actual value of bytesPerRow, defaulting to `align(minBytesPerRow, kBytesPerRowAlignment}`
+   * if not overridden.
+   */
   bytesPerRow: number;
+  /** Actual value of rowsPerImage, defaulting to `mipSize[1]` if not overridden. */
   rowsPerImage: number;
+  /** Size of the mip level being copied. */
   mipSize: [number, number, number];
 }
 
+/**
+ * Computes layout information for a copy of size `size` to/from a GPUTexture with the provided
+ * `format` and `dimension`.
+ *
+ * Computes default values for `bytesPerRow` and `rowsPerImage` if not specified.
+ */
 export function getTextureCopyLayout(
   format: SizedTextureFormat,
   dimension: GPUTextureDimension,
@@ -64,7 +65,7 @@ export function getTextureCopyLayout(
   const alignedMinBytesPerRow = align(minBytesPerRow, kBytesPerRowAlignment);
   if (bytesPerRow !== undefined) {
     assert(bytesPerRow >= alignedMinBytesPerRow);
-    assert(isAligned(bytesPerRow, kBytesPerRowAlignment));
+    assert(bytesPerRow % kBytesPerRowAlignment === 0);
   } else {
     bytesPerRow = alignedMinBytesPerRow;
   }
@@ -90,6 +91,13 @@ export function getTextureCopyLayout(
   };
 }
 
+/**
+ * Fill an ArrayBuffer with the linear-memory representation of a solid-color
+ * texture where every texel has the byte value `texelValue`.
+ * Preserves the contents of `outputBuffer` which are in "padding" space between image rows.
+ *
+ * Effectively emulates a copyTextureToBuffer from a solid-color texture to a buffer.
+ */
 export function fillTextureDataWithTexelValue(
   texelValue: ArrayBuffer,
   format: SizedTextureFormat,
@@ -125,6 +133,10 @@ export function fillTextureDataWithTexelValue(
   }
 }
 
+/**
+ * Create a `COPY_SRC` GPUBuffer containing the linear-memory representation of a solid-color
+ * texture where every texel has the byte value `texelValue`.
+ */
 export function createTextureUploadBuffer(
   texelValue: ArrayBuffer,
   device: GPUDevice,
@@ -160,4 +172,143 @@ export function createTextureUploadBuffer(
     bytesPerRow,
     rowsPerImage,
   };
+}
+
+export type ImageCopyType = 'WriteTexture' | 'CopyB2T' | 'CopyT2B';
+export const kImageCopyTypes: readonly ImageCopyType[] = [
+  'WriteTexture',
+  'CopyB2T',
+  'CopyT2B',
+] as const;
+
+/**
+ * Computes `bytesInACompleteRow` (as defined by the WebGPU spec) for image copies (B2T/T2B/writeTexture).
+ */
+export function bytesInACompleteRow(copyWidth: number, format: SizedTextureFormat): number {
+  const info = kSizedTextureFormatInfo[format];
+  assert(copyWidth % info.blockWidth === 0);
+  return (info.bytesPerBlock * copyWidth) / info.blockWidth;
+}
+
+function validateBytesPerRow({
+  bytesPerRow,
+  bytesInLastRow,
+  sizeInBlocks,
+}: {
+  bytesPerRow: number | undefined;
+  bytesInLastRow: number;
+  sizeInBlocks: Required<GPUExtent3DDict>;
+}) {
+  // If specified, layout.bytesPerRow must be greater than or equal to bytesInLastRow.
+  if (bytesPerRow !== undefined && bytesPerRow < bytesInLastRow) {
+    return false;
+  }
+  // If heightInBlocks > 1, layout.bytesPerRow must be specified.
+  // If copyExtent.depthOrArrayLayers > 1, layout.bytesPerRow and layout.rowsPerImage must be specified.
+  if (
+    bytesPerRow === undefined &&
+    (sizeInBlocks.height > 1 || sizeInBlocks.depthOrArrayLayers > 1)
+  ) {
+    return false;
+  }
+  return true;
+}
+
+function validateRowsPerImage({
+  rowsPerImage,
+  sizeInBlocks,
+}: {
+  rowsPerImage: number | undefined;
+  sizeInBlocks: Required<GPUExtent3DDict>;
+}) {
+  // If specified, layout.rowsPerImage must be greater than or equal to heightInBlocks.
+  if (rowsPerImage !== undefined && rowsPerImage < sizeInBlocks.height) {
+    return false;
+  }
+  // If copyExtent.depthOrArrayLayers > 1, layout.bytesPerRow and layout.rowsPerImage must be specified.
+  if (rowsPerImage === undefined && sizeInBlocks.depthOrArrayLayers > 1) {
+    return false;
+  }
+  return true;
+}
+
+interface DataBytesForCopyArgs {
+  layout: GPUImageDataLayout;
+  format: SizedTextureFormat;
+  copySize: Readonly<GPUExtent3DDict> | readonly number[];
+  method: ImageCopyType;
+}
+
+/**
+ * Validate a copy and compute the number of bytes it needs. Throws if the copy is invalid.
+ */
+export function dataBytesForCopyOrFail(args: DataBytesForCopyArgs): number {
+  const { minDataSizeOrOverestimate, copyValid } = dataBytesForCopyOrOverestimate(args);
+  assert(copyValid, 'copy was invalid');
+  return minDataSizeOrOverestimate;
+}
+
+/**
+ * Validate a copy and compute the number of bytes it needs. If the copy is invalid, attempts to
+ * "conservatively guess" (overestimate) the number of bytes that could be needed for a copy, even
+ * if the copy parameters turn out to be invalid. This hopes to avoid "buffer too small" validation
+ * errors when attempting to test other validation errors.
+ */
+export function dataBytesForCopyOrOverestimate({
+  layout,
+  format,
+  copySize: copySize_,
+  method,
+}: DataBytesForCopyArgs): { minDataSizeOrOverestimate: number; copyValid: boolean } {
+  const copyExtent = standardizeExtent3D(copySize_);
+
+  const info = kSizedTextureFormatInfo[format];
+  assert(copyExtent.width % info.blockWidth === 0);
+  assert(copyExtent.height % info.blockHeight === 0);
+  const sizeInBlocks = {
+    width: copyExtent.width / info.blockWidth,
+    height: copyExtent.height / info.blockHeight,
+    depthOrArrayLayers: copyExtent.depthOrArrayLayers,
+  } as const;
+  const bytesInLastRow = sizeInBlocks.width * info.bytesPerBlock;
+
+  let valid = true;
+  const offset = layout.offset ?? 0;
+  if (method !== 'WriteTexture') {
+    if (offset % info.bytesPerBlock !== 0) valid = false;
+    if (layout.bytesPerRow && layout.bytesPerRow % 256 !== 0) valid = false;
+  }
+
+  let requiredBytesInCopy = 0;
+  {
+    let { bytesPerRow, rowsPerImage } = layout;
+
+    // If bytesPerRow or rowsPerImage is invalid, guess a value for the sake of various tests that
+    // don't actually care about the exact value.
+    // (In particular for validation tests that want to test invalid bytesPerRow or rowsPerImage but
+    // need to make sure the total buffer size is still big enough.)
+    if (!validateBytesPerRow({ bytesPerRow, bytesInLastRow, sizeInBlocks })) {
+      bytesPerRow = undefined;
+      valid = false;
+    }
+    if (!validateRowsPerImage({ rowsPerImage, sizeInBlocks })) {
+      rowsPerImage = undefined;
+      valid = false;
+    }
+    // Pick values for cases when (a) bpr/rpi was invalid or (b) they're validly undefined.
+    bytesPerRow ??= align(info.bytesPerBlock * sizeInBlocks.width, 256);
+    rowsPerImage ??= sizeInBlocks.height;
+
+    if (copyExtent.depthOrArrayLayers > 1) {
+      const bytesPerImage = bytesPerRow * rowsPerImage;
+      const bytesBeforeLastImage = bytesPerImage * (copyExtent.depthOrArrayLayers - 1);
+      requiredBytesInCopy += bytesBeforeLastImage;
+    }
+    if (copyExtent.depthOrArrayLayers > 0) {
+      if (sizeInBlocks.height > 1) requiredBytesInCopy += bytesPerRow * (sizeInBlocks.height - 1);
+      if (sizeInBlocks.height > 0) requiredBytesInCopy += bytesInLastRow;
+    }
+  }
+
+  return { minDataSizeOrOverestimate: offset + requiredBytesInCopy, copyValid: valid };
 }
