+++ src/webgpu/util/texture/subresource.ts
@@ -1,12 +1,10 @@
-import { assert } from '../../../common/util/util.js';
-import { kAllTextureFormatInfo } from '../../capability_info.js';
-import { align } from '../../util/math.js';
-
+/** A range of indices expressed as { begin, count }. */
 export interface BeginCountRange {
   begin: number;
   count: number;
 }
 
+/* A range of indices, expressed as { begin, end }. */
 export interface BeginEndRange {
   begin: number;
   end: number;
@@ -22,6 +20,10 @@ function* rangeAsIterator(r: BeginEndRange | BeginCountRange): Generator<number>
   }
 }
 
+/**
+ * Represents a range of subresources of a single-plane texture:
+ * a min/max mip level and min/max array layer.
+ */
 export class SubresourceRange {
   readonly mipRange: BeginEndRange;
   readonly layerRange: BeginEndRange;
@@ -40,6 +42,9 @@ export class SubresourceRange {
     };
   }
 
+  /**
+   * Iterates over the "rectangle" of { mip level, array layer } pairs represented by the range.
+   */
   *each(): Generator<{ level: number; layer: number }> {
     for (let level = this.mipRange.begin; level < this.mipRange.end; ++level) {
       for (let layer = this.layerRange.begin; layer < this.layerRange.end; ++layer) {
@@ -48,6 +53,10 @@ export class SubresourceRange {
     }
   }
 
+  /**
+   * Iterates over the mip levels represented by the range, each level including an iterator
+   * over the array layers at that level.
+   */
   *mipLevels(): Generator<{ level: number; layers: Generator<number> }> {
     for (let level = this.mipRange.begin; level < this.mipRange.end; ++level) {
       yield {
@@ -56,28 +65,4 @@ export class SubresourceRange {
       };
     }
   }
-}
-
-// TODO(jiawei.shao@intel.com): support 1D and 3D textures
-export function physicalMipSize(
-  size: Required<GPUExtent3DDict>,
-  format: GPUTextureFormat,
-  dimension: GPUTextureDimension,
-  level: number
-): Required<GPUExtent3DDict> {
-  assert(dimension === '2d');
-  assert(Math.max(size.width, size.height) >> level > 0);
-
-  const virtualWidthAtLevel = Math.max(size.width >> level, 1);
-  const virtualHeightAtLevel = Math.max(size.height >> level, 1);
-  const physicalWidthAtLevel = align(virtualWidthAtLevel, kAllTextureFormatInfo[format].blockWidth);
-  const physicalHeightAtLevel = align(
-    virtualHeightAtLevel,
-    kAllTextureFormatInfo[format].blockHeight
-  );
-  return {
-    width: physicalWidthAtLevel,
-    height: physicalHeightAtLevel,
-    depthOrArrayLayers: size.depthOrArrayLayers,
-  };
 }
